requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.4,<3.0.0
python-multipart==0.0.6

# HTTP requests
requests>=2.32.5,<3.0.0
httpx==0.25.2

# Google APIs
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.25.2

# AI and LangChain
langchain==0.3.27
langchain-community==0.3.30
langchain-openai==0.3.34
openai==2.0.1
tiktoken==0.11.0
pypdf==6.1.1
# chromadb==0.4.18  # Commented out due to compilation issues

# Data processing
python-dateutil==2.8.2
pytz==2023.3

# Environment and configuration
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: For better JSON handling
orjson>=3.9.14,<4.0.0

# Optional: For email parsing
email-validator==2.1.0

# Optional: For natural language processing
nltk==3.8.1
spacy==3.7.2

# Optional: For calendar parsing
dateparser==1.1.8
